Added set index function

+++ src-tauri/src/app_state.rs
@@ -4,7 +4,7 @@ use std::sync::Mutex;
 pub struct InteriorAppState {
     pub db: SqlitePool,
     pub current_playlist: Vec<String>,
-    pub current_playlist_idx: i32,
+    pub current_playlist_idx: usize,
 }
 pub struct AppState {
     pub state: Mutex<InteriorAppState>,
